chore: renaming files
asset-interface-description.ts --> asset-interfaces-description.ts
AssetInterfaceDescriptionTest  --> asset-interfaces-description-test.ts
