+++ node/aas-aid/src/asset-interfaces-description.ts
